huffman_tree/requirements.txt
PrettyPrintTree == 1.36.2